Make LibLoad itself into a library
This was done to allow programs to include a dependency on a minimum
version of LibLoad, just like other library dependencies. It also allows
potentially exporting functions from LibLoad itself.

Because LibLoad has no internal replacements, due to its own current
implementation, it will not waste RAM space appending itself to the end
of a running program.

To go along with this change, LibLoad's version has been incremented to
version 3.0.